requirements.txt
feedparser==6.0.11
trafilatura==1.12.2
sumy==0.11.0
rapidfuzz==3.9.6
python-dateutil==2.9.0.post0
requests==2.32.3
openai==1.51.0
